==> src/1.2_YT_SB_Comments_Extraction.ipynb <==
# Extracting YouTube Comments
Video ID : https://www.youtube.com/watch?v=bFEoMO0pc7k

!pip install --upgrade google-api-python-client

from googleapiclient.discovery import build
import pandas as pd
import os

**Original Comments**

# Scrape Comments for SQL Using Python Through The Youtube Data API
api_key = 'AIzaSyCTQiMLrm5vTfNCiJSWATGbsehuwAOzezM' # Insert your Api key here.

youtube = build('youtube', 'v3', developerKey=api_key)

box = [['name', 'comment', 'date', 'likes', 'reply_count', "parent_id", "is_parent"]]

code_lang = [{"id":"bFEoMO0pc7k"}]

for id_code in code_lang:
  data = youtube.commentThreads().list(part='snippet', videoId=id_code['id'], maxResults='100', textFormat="plainText").execute()
  for i in data["items"]:
    name = i["snippet"]['topLevelComment']["snippet"]["authorDisplayName"]
    comment = i["snippet"]['topLevelComment']["snippet"]["textDisplay"]
    published_at = i["snippet"]['topLevelComment']["snippet"]['publishedAt']
    likes = i["snippet"]['topLevelComment']["snippet"]['likeCount']
    parent = i["snippet"]['topLevelComment']["id"]
    totalReplyCount = i["snippet"]['totalReplyCount']
    
    box.append([name, comment, published_at, likes, totalReplyCount, parent, "Yes" ])          
                        
    while ("nextPageToken" in data):
      data = youtube.commentThreads().list(part='snippet', videoId=id_code['id'], pageToken=data["nextPageToken"],
                                             maxResults='100', textFormat="plainText").execute()
                                             
      for i in data["items"]:
        name = i["snippet"]['topLevelComment']["snippet"]["authorDisplayName"]
        comment = i["snippet"]['topLevelComment']["snippet"]["textDisplay"]
        published_at = i["snippet"]['topLevelComment']["snippet"]['publishedAt']
        likes = i["snippet"]['topLevelComment']["snippet"]['likeCount']
        totalReplyCount = i["snippet"]['totalReplyCount']
        parent = i["snippet"]['topLevelComment']["id"]

        box.append([name, comment, published_at, likes, totalReplyCount, parent, "Yes" ])
                

df = pd.DataFrame({'name': [i[0] for i in box], 'comment': [i[1] for i in box], 'published_at': [i[2] for i in box],
                       'likes': [i[3] for i in box], 'Reply Count': [i[4] for i in box], 'parent_id': [i[5] for i in box], 'is_parent': [i[6] for i in box]})      
sql_vids = pd.DataFrame([])
sql_vids = sql_vids.append(df, ignore_index = True)

sql_vids_parents = sql_vids.drop(0, axis=0)
sql_vids_parents['Reply Count'] = sql_vids_parents['Reply Count'].map(int)

# Filter only comments with replies
sql_vids_parents_replies = sql_vids_parents[sql_vids_parents["Reply Count"]>0]

# of comments with replies
print(len(sql_vids_parents_replies))
sql_vids_parents_replies.head()

**Replies only**

api_key = "" # Insert your Api key here.

youtube = build('youtube', 'v3', developerKey=api_key)

box = [['name', 'comment', 'date', 'likes', 'reply_count', "parent_id", "is_parent"]]

for parent_id in sql_vids_parents_replies['parent_id']:
  data = youtube.comments().list(part='snippet', maxResults='100', textFormat="plainText", parentId=parent_id).execute()
  for i in data["items"]:
    name = i["snippet"]["authorDisplayName"]
    comment = i["snippet"]["textDisplay"]
    published_at = i["snippet"]['publishedAt']
    likes = i["snippet"]['likeCount']
    parent = parent_id
    totalReplyCount = ''
    
    box.append([name, comment, published_at, likes, totalReplyCount, parent, "No" ])          
                        
    while ("nextPageToken" in data):
      data = youtube.comments().list(part='snippet', pageToken=data["nextPageToken"],
                                             maxResults='100', textFormat="plainText", parentId=parent_id).execute()
                                             
      for i in data["items"]:
        name = i["snippet"]["authorDisplayName"]
        comment = i["snippet"]["textDisplay"]
        published_at = i["snippet"]['publishedAt']
        likes = i["snippet"]['likeCount']
        parent = parent_id
        totalReplyCount = ''

        box.append([name, comment, published_at, likes, totalReplyCount, parent, "No" ])
                

df_replies = pd.DataFrame({'name': [i[0] for i in box], 'comment': [i[1] for i in box], 'published_at': [i[2] for i in box],
                       'likes': [i[3] for i in box], 'Reply Count': [i[4] for i in box], 'parent_id': [i[5] for i in box], 'is_parent': [i[6] for i in box]})      
sql_vids_replies = pd.DataFrame([])
sql_vids_replies = sql_vids_replies.append(df_replies, ignore_index = True)

sql_vids_replies.drop(0, axis=0, inplace=True)

sql_vids_replies.head()

**Combining the datasets**

df_all = pd.concat([sql_vids, sql_vids_replies])

# Total # of records
len(df_all)

# Saving files
df_all.to_csv(os.path.join('../output/','06112022_SB_youtube_comments.csv'),encoding='utf_8_sig')
df_all.to_pickle(os.path.join('../output/','06112022_SB_youtube_comments.pkl'))
