==> src/3.1_YT_All_Comments_Topic_Model.ipynb <==
import pandas as pd
import numpy as np
import os
from gensim.models import word2vec

df_comments = pd.read_pickle('../output/07142022_clean_data.pkl')
df_comments.head()

lemmaSenetence = df_comments.lemmasNgrams.values
model = word2vec.Word2Vec(lemmaSenetence, vector_size=50, min_count=3, epochs=20)

topicSubject =['black','people','violence','rape','ridicule','feminist','critique', 'gay', 'musicality','comic','men']
topicBlack = model.wv.most_similar('black', topn=50)
topicPeople = model.wv.most_similar('people', topn=50)
topicViolence = model.wv.most_similar('violence', topn=50)
topicRape = model.wv.most_similar('rape', topn=50)
topicRidicule = model.wv.most_similar('ridicule', topn=50)
topicFeminist = model.wv.most_similar('feminist', topn=50)
topicCritique = model.wv.most_similar('criticize', topn=50)
topicGay = model.wv.most_similar('gay', topn=50)
topicMusicality = model.wv.most_similar('musicality', topn=50)
topicComic = model.wv.most_similar('comic', topn=50)
topicMen = model.wv.most_similar('men', topn=50)


model.wv.most_similar('black_people', topn=50)

topicCols =['topicBlack','topicPeople','topicViolence','topicRape','topicRidicule','topicFeminist','topicCritique', 'topicGay', 'topicMusicality','topicComic','topicMen']
for col in topicCols:
    df_comments[col] = None

topicColsScore = ['topicBlackS','topicPeopleS','topicViolenceS','topicRapeS','topicRidiculeS','topicFeministS','topicCritiqueS', 'topicGayS', 'topicMusicalityS','topicComicS','topicMenS']
for col in topicColsScore:
    df_comments[col] = None

'Exporting top 10 words in each topic to a csv'
df=pd.DataFrame()
df['topic']=''
df['words']=''
j=0
for col in topicCols:
    i=0
    for items in eval(col):
        #print(items)
        df.loc[j,'topic'] = col
        df.at[j,'words'] = items
        i=i+1
        j=j+1
        if i ==10:
            break
df.to_csv(os.path.join('../output/','topic_list_07202022.csv'))

def sim_word(topicName):
    "Extracting only words from similarity output of wordvec"
    simList = set([w[0] for w in topicName if w[1]>.5])
    return simList

def sim_score(topicName):
    "Extracting word with score from similarity output of wordvec"
    simScore = set([w for w in topicName if w[1]>.5])
    return simScore

"Putting all the sim words in one dictionary by topic"
topicDict = {}
scoreDict = {}
i=0
for col in topicCols:
    topicDict.update({col: sim_word(eval(col))})
    topicDict[col].add(topicSubject[i])
    scoreDict.update({col: sim_score(eval(col))})
    scoreDict[col].add((topicSubject[i],1))
    i=i+1

'Sample checks'
n=2
v=df_comments['lemmasNgrams'][n]
#v= ['work', 'screen', 'dirty', 'look_like', 'phone', 'drop', 'coffee', 'stain', 'spot']
for topic_, sim_ in topicDict.items():
    if len(set(sim_).intersection(set(v)))>0:
        common =(set(sim_).intersection(set(v)))
        print (topic_, common)
        cosScore=0
        for items in scoreDict[topic_]:
            # Avg. Similarity score
            if items[0] in common:
                print(items[1])
                cosScore = items[1] + cosScore
        cosScoreAvg =  cosScore/len(common)
        assert cosScoreAvg <= 1
        print(cosScoreAvg)

print(df_comments['comment'][n])

'Finding common words between lemmas of review text and most similar words by each topic'
df_comments['allTopics']=''
for index, row in df_comments.iterrows():
    lemmaCheck = row['lemmasNgrams']
    for topic_, sim_ in topicDict.items():
        common = set(sim_).intersection(set(lemmaCheck))
        if len(common)>0:
            df_comments.at[index, topic_] = common
            df_comments.loc[index, 'allTopics'] = topic_ + ', ' + df_comments.loc[index, 'allTopics']
            cosScore=0  # Cosine similarity score
            for items in scoreDict[topic_]:
                # Avg. Similarity score
                if items[0] in common:
                    # print(items[1])
                    cosScore = items[1] + cosScore
            cosScoreAvg =  cosScore/len(common) # Avg. cosine similarity score if there are multiple common words
            assert cosScoreAvg <= 1
            df_comments.loc[index, (topic_+'S')] = cosScoreAvg


df_comments.groupby('allTopics')['parent_id'].count().sort_values(ascending=False).head(12)

df_comments.to_pickle(os.path.join('../output/','all_topic_output_07202022.pkl'))

df_comments.to_csv(os.path.join('../output/','all_topic_output_07202022.csv'),encoding='utf_8_sig')
