==> src/3.2_YT_Comments_Sentiment_Analysis.ipynb <==
!pip install vaderSentiment

import pandas as pd
import os
import numpy as np
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

df = pd.read_excel('../output/07142022_clean_data.xlsx')

df.head()

analyzer = SentimentIntensityAnalyzer()
df['clean_comments'] = df['clean_comments'].map(str)
df['sentiment'] = df['clean_comments'].apply(lambda x: analyzer.polarity_scores(x))
df['compound'] = df['sentiment'].apply(lambda score_dict: score_dict['compound'])
df['positive'] = df['sentiment'].apply(lambda score_dict: score_dict['pos'])
df['negative'] = df['sentiment'].apply(lambda score_dict: score_dict['neg'])
df['neutral'] = df['sentiment'].apply(lambda score_dict: score_dict['neu'])

# Assigning labels (positive, neutral and negative sentiments)
df.loc[df.compound > .5, 'sentiment_type'] = 'POSITIVE'
df.loc[(df.compound >= -.5) & (df.compound <= .5), 'sentiment_type'] = 'NEUTRAL'
df.loc[df.compound < -.5, 'sentiment_type'] = 'NEGATIVE'

df.head()

df.value_counts('sentiment_type')

df.to_excel('../output/07202022_clean_data.xlsx')
