==> src/4.2_Key_Emotions_Extraction_AO.ipynb <==
pip install --ignore-installed --upgrade tensorflow-gpu

import sys
!{sys.executable} -m pip install transformers

from transformers import pipeline
classifier = pipeline("text-classification", model="j-hartmann/emotion-english-distilroberta-base", return_all_scores=True)

import pandas as pd
import os

df_comments = pd.read_pickle('07142022_clean_data.pkl')

df_comments.video_link.unique()

df = df_comments[df_comments.video_link == 'Annoying Orange']

print(df.shape)
df = df.reset_index(drop=True)

import torch
import pandas as pd
import numpy as np
from transformers import AutoTokenizer, AutoModelForSequenceClassification, Trainer

# Create class for data preparation
class SimpleDataset:
    def __init__(self, tokenized_texts):
        self.tokenized_texts = tokenized_texts

    def __len__(self):
        return len(self.tokenized_texts["input_ids"])

    def __getitem__(self, idx):
        return {k: v[idx] for k, v in self.tokenized_texts.items()}

pred_texts = df['clean_comments'].astype('str').tolist()

# load tokenizer and model, create trainer
model_name = "j-hartmann/emotion-english-distilroberta-base"
tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModelForSequenceClassification.from_pretrained(model_name)
trainer = Trainer(model=model)

# Tokenize texts and create prediction data set
tokenized_texts = tokenizer(pred_texts,truncation=True,padding=True)
pred_dataset = SimpleDataset(tokenized_texts)

# Run predictions
predictions = trainer.predict(pred_dataset)

# Transform predictions to labels
preds = predictions.predictions.argmax(-1)
labels = pd.Series(preds).map(model.config.id2label)
scores = (np.exp(predictions[0])/np.exp(predictions[0]).sum(-1,keepdims=True)).max(1)

# scores raw
temp = (np.exp(predictions[0])/np.exp(predictions[0]).sum(-1,keepdims=True))

df['anger']= None
df['disgust']= None
df['fear']= None
df['joy']= None
df['neutral']= None
df['sadness']= None
df['surprise']= None

for index, row in df.iterrows():
    df.loc[index, 'anger']= temp[index][0]
    df.loc[index, 'disgust']= temp[index][1]
    df.loc[index, 'fear']= temp[index][2]
    df.loc[index, 'joy']= temp[index][3]
    df.loc[index, 'neutral']= temp[index][4]
    df.loc[index, 'sadness']= temp[index][5]
    df.loc[index, 'surprise']= temp[index][6]
    print(index)

df.to_pickle('Annoying_Orange_Comments_Key_Emotions.pkl')
