==> src/5.1_Toxicity_Parameter_Generation.ipynb <==
import pandas as pd
import os
from detoxify import Detoxify

df = pd.read_pickle('07142022_clean_data.pkl')

df['toxicity']= None
df['severe_toxicity']= None
df['obscene']= None
df['identity_attack']= None
df['insult']= None
df['sexual_explicit']= None
df['threat']= None
df['toxicity_check'] = None

for index, row in df.iterrows():
    toxic = Detoxify('unbiased').predict(df.loc[index,'clean_comments'])
    df.at[index, 'toxicity_check'] = toxic
    df.loc[index, 'toxicity']= toxic.get('toxicity',0)
    df.loc[index, 'severe_toxicity']= toxic.get('severe_toxicity', 0)
    df.loc[index, 'obscene']= toxic.get('obscene', 0)
    df.loc[index, 'identity_attack']= toxic.get('identity_attack', 0)
    df.loc[index, 'insult']= toxic.get('insult', 0)
    df.loc[index, 'sexual_explicit']= toxic.get('sexual_explicit', 0)
    df.loc[index, 'threat']= toxic.get('threat', 0)


df.to_pickle('Final_Combined_Df_Detoxify_Results.pkl')
