==> src/6.1_Combine_All_Emotion_Extraction.ipynb <==
# Install a pip package in the current Jupyter kernel
import sys
!{sys.executable} -m pip install pickle5

import pandas as pd
import os
import pickle5 as pickle

import sys
!{sys.executable} -m pip install pandas==1.4.1

#### Detoxify Output

final_df = pd.read_pickle('Final_Combined_Df_Detoxify_Results.pkl')

final_df.head()

### Emotion Output

df1 = pd.read_pickle('Bed_Intruder_Comments_Key_Emotions.pkl')
df2 = pd.read_pickle('Sweet_Brown_Comments_Key_Emotions.pkl')
df3 = pd.read_pickle('Annoying_Orange_Comments_Key_Emotions.pkl')
df4 = pd.read_pickle('Justice_League_Comments_Key_Emotions.pkl')
df5 = pd.read_pickle('World_Star_Comments_Key_Emotions.pkl')

all_emotions = pd.concat([df1,df2,df3,df4,df5])
all_emotions.reset_index(drop=True, inplace =True)

print(len(all_emotions))

all_emotions.head()

### Merge to Check if all comments are same row by row

all_combined= final_df.join([all_emotions])

final_df['anger'] = None
final_df['disgust'] = None
final_df['fear'] = None
final_df['joy'] = None
final_df['neutral'] = None
final_df['sadness'] = None
final_df['surprise'] = None
final_df['comments_new'] = None

for index, row in final_df.iterrows():
    if final_df.loc[index,'comment'] != all_emotions.loc[index,'comment']:
        print(index)
        break
    else:
        final_df.loc[index,'anger']= all_emotions.loc[index,'anger']
        final_df.loc[index,'disgust']= all_emotions.loc[index,'disgust']
        final_df.loc[index,'fear']=all_emotions.loc[index,'fear']
        final_df.loc[index,'joy']=all_emotions.loc[index,'joy']
        final_df.loc[index,'neutral']=all_emotions.loc[index,'neutral']
        final_df.loc[index,'sadness']=all_emotions.loc[index,'sadness']
        final_df.loc[index,'surprise']= all_emotions.loc[index,'surprise']

final_df.to_pickle('07212022_All_Processed_Comments.pkl') # Flows into Tableau report
